Separate CDC routing from metadata mutation application
