+++ fdbserver/logsystem/include/fdbserver/logsystem/ApplyMetadataMutation.h
@@ -39,6 +39,7 @@
 #include "flow/FastRef.h"
 
 class AccumulativeChecksumBuilder;
+class CDCRoutingTable;
 
 // applyMetadataMutations() is shared with resolver/clustercontroller code, so
 // it depends on a narrow commit-proxy view instead of the full ProxyCommitData
@@ -54,31 +55,6 @@ struct ApplyMutationsData {
 	Future<Void> worker;
 	Version endVersion;
 	Reference<KeyRangeMap<Version>> keyVersion;
-};
-
-// Active CDC write routing reconstructed from durable stream and tag-history metadata.
-class CDCRoutingTable : NonCopyable {
-	struct StreamState {
-		Optional<KeyRange> keys;
-		Optional<std::pair<Version, Tag>> tag;
-	};
-
-	std::unordered_map<CDCStreamId, StreamState> streams;
-	KeyRangeMap<std::set<Tag>> tagsByRange;
-
-	void updateRange(CDCStreamId streamId, KeyRangeRef const& keys);
-	bool updateTag(CDCStreamId streamId, Version version, Tag tag);
-	void rebuildRanges();
-
-public:
-	CDCRoutingTable();
-	void setRange(CDCStreamId streamId, KeyRangeRef const& keys);
-	void setTag(CDCStreamId streamId, Version version, Tag tag);
-	void reload(IKeyValueStore* txnStateStore);
-	bool empty() const { return streams.empty(); }
-
-	const std::set<Tag>& tagsForKey(KeyRef const& key) const;
-	std::set<Tag> tagsForRange(KeyRangeRef const& keys) const;
 };
 
 struct ApplyMetadataProxyContext {
